08 Jan 2024

+++ js/user-mangment.js
@@ -12,13 +12,13 @@ function addUser() {
     var form = document.getElementById("addUserForm");
 
     var newRow = table.insertRow(table.rows.length);
+  
     var cell1 = newRow.insertCell(0);
     var cell2 = newRow.insertCell(1);
     var cell3 = newRow.insertCell(2);
     var cell4 = newRow.insertCell(3);
     var cell5 = newRow.insertCell(4);
     var cell6 = newRow.insertCell(5);
-    var cell7 = newRow.insertCell(6);
 
     var firstName = form.elements["firstName"].value;
     var empID = form.elements["empID"].value;
@@ -26,15 +26,20 @@ function addUser() {
     var email = form.elements["email"].value;
     var role = form.elements["role"].value;
 
-    cell1.innerHTML = '<input type="checkbox">';
-    cell2.innerHTML = empID;
-    cell3.innerHTML = firstName;
-    cell4.innerHTML = lastName;
-    cell5.innerHTML = email;
-    cell6.innerHTML = role;
+    if (!firstName || !lastName || !email || !role) {
+        alert("Please fill in all required fields.");
+        return;
+    }
+
+   // cell1.innerHTML = '<input type="checkbox">';
+    cell1.innerHTML = empID;
+    cell2.innerHTML = firstName;
+    cell3.innerHTML = lastName;
+    cell4.innerHTML = email;
+    cell5.innerHTML = role;
 
     // Action buttons
-    cell7.innerHTML = '<button class="edit-btn" onclick="editUser(this)"><i class="fa fa-edit"></i></button>' +
+    cell6.innerHTML = '<button class="edit-btn" onclick="editUser(this)"><i class="fa fa-edit"></i></button>' +
         '<button class="delete-btn" onclick="deleteUser(this)"><i class="fa fa-trash"></i></button>';
 
     // Clear the form fields
